订单状态定时处理

+++ sky-pojo/src/main/java/com/sky/dto/OrdersCancelDTO.java
@@ -1,5 +1,6 @@
 package com.sky.dto;
 
+import io.swagger.models.auth.In;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -7,7 +8,7 @@ import java.io.Serializable;
 @Data
 public class OrdersCancelDTO implements Serializable {
 
-    private Long id;
+    private Integer id;
     //订单取消原因
     private String cancelReason;
