+++ sky-pojo/src/main/java/com/sky/dto/OrdersRejectionDTO.java
@@ -1,5 +1,6 @@
 package com.sky.dto;
 
+import io.swagger.models.auth.In;
 import lombok.Data;
 
 import java.io.Serializable;
@@ -7,7 +8,7 @@ import java.io.Serializable;
 @Data
 public class OrdersRejectionDTO implements Serializable {
 
-    private Long id;
+    private Integer id;
 
     //订单拒绝原因
     private String rejectionReason;
